requirements.txt
fastapi
pydantic==1.10.10
python-dotenv==1.0.0
Requests==2.31.0
uvicorn==0.23.2
validators==0.20.0
pyresparser==1.0.6
gunicorn==21.2.0
spacy==3.7.2
nltk==3.8.1
# fastapi_session==0.1.1
